fix: wait for transaction
